+++ aliding-20230426/alibabacloud_aliding20230426/models/_batch_query_group_member_response_body.py
@@ -1,0 +1,75 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from typing import List
+
+from darabonba.model import DaraModel
+
+class BatchQueryGroupMemberResponseBody(DaraModel):
+    def __init__(
+        self,
+        has_more: bool = None,
+        member_user_ids: List[str] = None,
+        next_token: str = None,
+        request_id: str = None,
+        vendor_request_id: str = None,
+        vendor_type: str = None,
+    ):
+        self.has_more = has_more
+        self.member_user_ids = member_user_ids
+        self.next_token = next_token
+        self.request_id = request_id
+        self.vendor_request_id = vendor_request_id
+        self.vendor_type = vendor_type
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.has_more is not None:
+            result['hasMore'] = self.has_more
+
+        if self.member_user_ids is not None:
+            result['memberUserIds'] = self.member_user_ids
+
+        if self.next_token is not None:
+            result['nextToken'] = self.next_token
+
+        if self.request_id is not None:
+            result['requestId'] = self.request_id
+
+        if self.vendor_request_id is not None:
+            result['vendorRequestId'] = self.vendor_request_id
+
+        if self.vendor_type is not None:
+            result['vendorType'] = self.vendor_type
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('hasMore') is not None:
+            self.has_more = m.get('hasMore')
+
+        if m.get('memberUserIds') is not None:
+            self.member_user_ids = m.get('memberUserIds')
+
+        if m.get('nextToken') is not None:
+            self.next_token = m.get('nextToken')
+
+        if m.get('requestId') is not None:
+            self.request_id = m.get('requestId')
+
+        if m.get('vendorRequestId') is not None:
+            self.vendor_request_id = m.get('vendorRequestId')
+
+        if m.get('vendorType') is not None:
+            self.vendor_type = m.get('vendorType')
+
+        return self
+