+++ aliding-20230426/alibabacloud_aliding20230426/models/_batch_query_group_member_shrink_request.py
@@ -1,0 +1,67 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from darabonba.model import DaraModel
+
+class BatchQueryGroupMemberShrinkRequest(DaraModel):
+    def __init__(
+        self,
+        cool_app_code: str = None,
+        max_results: int = None,
+        next_token: str = None,
+        open_conversation_id: str = None,
+        tenant_context_shrink: str = None,
+    ):
+        self.cool_app_code = cool_app_code
+        # This parameter is required.
+        self.max_results = max_results
+        self.next_token = next_token
+        # This parameter is required.
+        self.open_conversation_id = open_conversation_id
+        self.tenant_context_shrink = tenant_context_shrink
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.cool_app_code is not None:
+            result['CoolAppCode'] = self.cool_app_code
+
+        if self.max_results is not None:
+            result['MaxResults'] = self.max_results
+
+        if self.next_token is not None:
+            result['NextToken'] = self.next_token
+
+        if self.open_conversation_id is not None:
+            result['OpenConversationId'] = self.open_conversation_id
+
+        if self.tenant_context_shrink is not None:
+            result['TenantContext'] = self.tenant_context_shrink
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('CoolAppCode') is not None:
+            self.cool_app_code = m.get('CoolAppCode')
+
+        if m.get('MaxResults') is not None:
+            self.max_results = m.get('MaxResults')
+
+        if m.get('NextToken') is not None:
+            self.next_token = m.get('NextToken')
+
+        if m.get('OpenConversationId') is not None:
+            self.open_conversation_id = m.get('OpenConversationId')
+
+        if m.get('TenantContext') is not None:
+            self.tenant_context_shrink = m.get('TenantContext')
+
+        return self
+