+++ aliding-20230426/alibabacloud_aliding20230426/models/_get_scencegroup_file_downloadurl_response_body.py
@@ -2,27 +2,18 @@
 # This file is auto-generated, don't edit it. Thanks.
 from __future__ import annotations
 
-from typing import List
-
 from darabonba.model import DaraModel
 
-class CreateScheduleConferenceResponseBody(DaraModel):
+class GetScencegroupFileDownloadurlResponseBody(DaraModel):
     def __init__(
         self,
-        phones: List[str] = None,
+        download_url: str = None,
         request_id: str = None,
-        room_code: str = None,
-        schedule_conference_id: str = None,
-        url: str = None,
         vendor_request_id: str = None,
         vendor_type: str = None,
     ):
-        self.phones = phones
-        # requestId
+        self.download_url = download_url
         self.request_id = request_id
-        self.room_code = room_code
-        self.schedule_conference_id = schedule_conference_id
-        self.url = url
         self.vendor_request_id = vendor_request_id
         self.vendor_type = vendor_type
 
@@ -34,20 +25,11 @@ class CreateScheduleConferenceResponseBody(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
-        if self.phones is not None:
-            result['phones'] = self.phones
+        if self.download_url is not None:
+            result['downloadUrl'] = self.download_url
 
         if self.request_id is not None:
             result['requestId'] = self.request_id
-
-        if self.room_code is not None:
-            result['roomCode'] = self.room_code
-
-        if self.schedule_conference_id is not None:
-            result['scheduleConferenceId'] = self.schedule_conference_id
-
-        if self.url is not None:
-            result['url'] = self.url
 
         if self.vendor_request_id is not None:
             result['vendorRequestId'] = self.vendor_request_id
@@ -59,20 +41,11 @@ class CreateScheduleConferenceResponseBody(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
-        if m.get('phones') is not None:
-            self.phones = m.get('phones')
+        if m.get('downloadUrl') is not None:
+            self.download_url = m.get('downloadUrl')
 
         if m.get('requestId') is not None:
             self.request_id = m.get('requestId')
-
-        if m.get('roomCode') is not None:
-            self.room_code = m.get('roomCode')
-
-        if m.get('scheduleConferenceId') is not None:
-            self.schedule_conference_id = m.get('scheduleConferenceId')
-
-        if m.get('url') is not None:
-            self.url = m.get('url')
 
         if m.get('vendorRequestId') is not None:
             self.vendor_request_id = m.get('vendorRequestId')